--- stacked_widget/main.cpp
@@ -1,11 +1,0 @@
-#include "widget.h"
-
-#include <QApplication>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-    Widget w;
-    w.show();
-    return a.exec();
-}
