--- stacked_widget/page_2players.ui
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>page_2players</class>
- <widget class="QWidget" name="page_2players">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>800</width>
-    <height>600</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <widget class="QPushButton" name="cheatButton">
-   <property name="geometry">
-    <rect>
-     <x>110</x>
-     <y>430</y>
-     <width>200</width>
-     <height>50</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>Cheat</string>
-   </property>
-   <property name="autoDefault">
-    <bool>false</bool>
-   </property>
-   <property name="default">
-    <bool>false</bool>
-   </property>
-   <property name="flat">
-    <bool>false</bool>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="cooperateButton">
-   <property name="geometry">
-    <rect>
-     <x>460</x>
-     <y>430</y>
-     <width>200</width>
-     <height>50</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>Cooperate</string>
-   </property>
-  </widget>
- </widget>
- <resources/>
- <connections/>
-</ui>
