--- stacked_widget/page_onetournament.ui
@@ -1,127 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>page_onetournament</class>
- <widget class="QWidget" name="page_onetournament">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>800</width>
-    <height>600</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <widget class="QTextBrowser" name="textBrowser">
-   <property name="geometry">
-    <rect>
-     <x>440</x>
-     <y>70</y>
-     <width>281</width>
-     <height>192</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">background-color: rgb(250, 250, 250);
-font: 9pt &quot;Segoe Print&quot;;</string>
-   </property>
-   <property name="frameShape">
-    <enum>QFrame::NoFrame</enum>
-   </property>
-   <property name="html">
-    <string>&lt;!DOCTYPE HTML PUBLIC &quot;-//W3C//DTD HTML 4.0//EN&quot; &quot;http://www.w3.org/TR/REC-html40/strict.dtd&quot;&gt;
-&lt;html&gt;&lt;head&gt;&lt;meta name=&quot;qrichtext&quot; content=&quot;1&quot; /&gt;&lt;meta charset=&quot;utf-8&quot; /&gt;&lt;style type=&quot;text/css&quot;&gt;
-p, li { white-space: pre-wrap; }
-hr { height: 1px; border-width: 0; }
-li.unchecked::marker { content: &quot;\2610&quot;; }
-li.checked::marker { content: &quot;\2612&quot;; }
-&lt;/style&gt;&lt;/head&gt;&lt;body style=&quot; font-family:'Segoe Print'; font-size:9pt; font-weight:400; font-style:normal;&quot;&gt;
-&lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;span style=&quot; font-family:'FuturaHandwritten'; font-size:20px; color:#333333; background-color:#ffffff;&quot;&gt;Each character will now play against every other character: that's 10 paired matches, and 10 rounds per match.&lt;/span&gt;&lt;/p&gt;
-&lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;span style=&quot; font-family:'FuturaHandwritten'; font-size:20px; color:#333333; background-color:#ffffff;&quot;&gt; Who do you think will get the highest total score? &lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-   </property>
-  </widget>
-  <widget class="QRadioButton" name="Copycat">
-   <property name="geometry">
-    <rect>
-     <x>460</x>
-     <y>270</y>
-     <width>299</width>
-     <height>41</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">font: 700 20pt &quot;Segoe Print&quot;;</string>
-   </property>
-   <property name="text">
-    <string>Copycat</string>
-   </property>
-  </widget>
-  <widget class="QRadioButton" name="AllCooperate">
-   <property name="geometry">
-    <rect>
-     <x>460</x>
-     <y>350</y>
-     <width>211</width>
-     <height>41</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">font: 700 20pt &quot;Segoe Print&quot;;</string>
-   </property>
-   <property name="text">
-    <string>All Cooperate</string>
-   </property>
-  </widget>
-  <widget class="QRadioButton" name="Grudger">
-   <property name="geometry">
-    <rect>
-     <x>460</x>
-     <y>390</y>
-     <width>211</width>
-     <height>41</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">font: 700 20pt &quot;Segoe Print&quot;;</string>
-   </property>
-   <property name="text">
-    <string>Grudger</string>
-   </property>
-  </widget>
-  <widget class="QRadioButton" name="Detective">
-   <property name="geometry">
-    <rect>
-     <x>460</x>
-     <y>430</y>
-     <width>191</width>
-     <height>41</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">font: 700 20pt &quot;Segoe Print&quot;;</string>
-   </property>
-   <property name="text">
-    <string>Detective</string>
-   </property>
-  </widget>
-  <widget class="QRadioButton" name="AllCheat">
-   <property name="geometry">
-    <rect>
-     <x>460</x>
-     <y>317</y>
-     <width>201</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">font: 700 20pt &quot;Segoe Print&quot;;</string>
-   </property>
-   <property name="text">
-    <string>All Cheat</string>
-   </property>
-  </widget>
- </widget>
- <resources/>
- <connections/>
-</ui>
