--- stacked_widget/widget.h
@@ -1,23 +1,0 @@
-#ifndef WIDGET_H
-#define WIDGET_H
-
-#include <QWidget>
-
-QT_BEGIN_NAMESPACE
-namespace Ui {
-class Widget;
-}
-QT_END_NAMESPACE
-
-class Widget : public QWidget
-{
-    Q_OBJECT
-
-public:
-    Widget(QWidget *parent = nullptr);
-    ~Widget();
-
-
-    Ui::Widget *ui;
-};
-#endif // WIDGET_H
