+++ widget.ui
@@ -1,0 +1,137 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ui version="4.0">
+ <class>Widget</class>
+ <widget class="QWidget" name="Widget">
+  <property name="geometry">
+   <rect>
+    <x>0</x>
+    <y>0</y>
+    <width>800</width>
+    <height>600</height>
+   </rect>
+  </property>
+  <property name="windowTitle">
+   <string>Widget</string>
+  </property>
+  <widget class="QStackedWidget" name="stackedWidget">
+   <property name="geometry">
+    <rect>
+     <x>20</x>
+     <y>20</y>
+     <width>760</width>
+     <height>500</height>
+    </rect>
+   </property>
+   <property name="frameShape">
+    <enum>QFrame::StyledPanel</enum>
+   </property>
+   <property name="currentIndex">
+    <number>3</number>
+   </property>
+   <widget class="page_welcome" name="page_1"/>
+   <widget class="page_rules" name="page_2"/>
+   <widget class="page_2players" name="page_3"/>
+   <widget class="page_onetournament" name="page_4"/>
+   <widget class="page_allplayers" name="page_5"/>
+  </widget>
+  <widget class="QPushButton" name="welcomeButton">
+   <property name="geometry">
+    <rect>
+     <x>60</x>
+     <y>540</y>
+     <width>50</width>
+     <height>50</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string/>
+   </property>
+  </widget>
+  <widget class="QPushButton" name="rulesButton">
+   <property name="geometry">
+    <rect>
+     <x>140</x>
+     <y>540</y>
+     <width>50</width>
+     <height>50</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string/>
+   </property>
+  </widget>
+  <widget class="QPushButton" name="twoplayerButton">
+   <property name="geometry">
+    <rect>
+     <x>220</x>
+     <y>540</y>
+     <width>50</width>
+     <height>50</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string/>
+   </property>
+  </widget>
+  <widget class="QPushButton" name="allplayerButton">
+   <property name="geometry">
+    <rect>
+     <x>380</x>
+     <y>540</y>
+     <width>50</width>
+     <height>50</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string/>
+   </property>
+  </widget>
+  <widget class="QPushButton" name="tournamentButton">
+   <property name="geometry">
+    <rect>
+     <x>300</x>
+     <y>540</y>
+     <width>50</width>
+     <height>50</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string/>
+   </property>
+  </widget>
+ </widget>
+ <customwidgets>
+  <customwidget>
+   <class>page_welcome</class>
+   <extends>QWidget</extends>
+   <header location="global">page_welcome.h</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>page_rules</class>
+   <extends>QWidget</extends>
+   <header location="global">page_rules.h</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>page_2players</class>
+   <extends>QWidget</extends>
+   <header location="global">page_2players.h</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>page_allplayers</class>
+   <extends>QWidget</extends>
+   <header location="global">page_allplayers.h</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>page_onetournament</class>
+   <extends>QWidget</extends>
+   <header location="global">page_onetournament.h</header>
+   <container>1</container>
+  </customwidget>
+ </customwidgets>
+ <resources/>
+ <connections/>
+</ui>
